Fixed some namespaces, move some using namescpaces to more readable order.
