+++ Lab1/Lab1/Program.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using CommandLine;
 using NLog;
+using Lab1.Models;
 using Lab1.Services.Exceptions;
 using Lab1.Services.Interfaces;
-using Lab1.Models;
 using Lab1.Services.IOSystem.Readers;
 using Lab1.Services.IOSystem.Writers;
