+++ Lab1/Lab1/Services/IOSystem/Readers/CsvReader.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Lab1.Services.Exceptions;
 using Lab1.Models;
+using Lab1.Services.Exceptions;
 using Lab1.Services.Interfaces;
 using Lab1.Services.Parsers;
 using Lab1.Services.Validation;
@@ -15,11 +15,9 @@ namespace Lab1.Services.IOSystem.Readers
         {
             path.CsvPathValidation();
             path.CheckFileExistance();
-            
             List<Student> students = new List<Student>();
             bool firstString = true;
             fieldNames = new List<string>();
-
             using (StreamReader streamReader = new StreamReader(path))
             {
                 while (!streamReader.EndOfStream)
@@ -40,11 +38,8 @@ namespace Lab1.Services.IOSystem.Readers
                         student.Validate();
                         students.Add(student);
                     }
-                }
-
-              
+                } 
             }
-
             return students;
         }
     }
